pyproject.toml
[tool.poetry]
name = "raggaeton"
version = "0.1.0"
description = "RAG and chat over your documents"
authors = ["Chen Enjiao <erniesg@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
requests = "^2.31.0"
pyyaml = "^6.0.1"
python-dotenv = "^1.0.1"
supabase = "^2.5.0"
modal = "^0.62.186"
llama-index = "^0.10.42"
llama-index-readers-database = "^0.1.4"
llama-index-embeddings-huggingface = "^0.2.1"
llama-index-vector-stores-supabase = "^0.1.5"
llama-index-llms-anthropic = "^0.1.11"
llama-index-tools-google = "^0.1.5"
llama-hub = "^0.0.79.post1"
llama-index-packs-ragatouille-retriever = "^0.1.3"
fastapi = "^0.111.0"
uvicorn = "^0.30.0"
ragatouille = "^0.0.8.post2"
google-cloud-secret-manager = "^2.20.0"
google-cloud-storage = "^2.16.0"
python-multipart = "^0.0.9"
aiohttp = "^3.9.5"
google-search-results = "^2.4.2"
serpapi = "^0.1.5"
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"
pytest = "^8.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"